refactor, delete preloader when showed modal error
