+++ src/js/utils_dev/temp-text.js
@@ -1,6 +1,5 @@
 'use strict';
 
-// Временно
 export const tempText = `Форма тела пингвинов обтекаемая, 
 что идеально для передвижения в воде. 
 Мускулатура и устройство костей позволяют им под водой работать
