+++ src/scss/modules/_radio-buttons.scss
@@ -7,16 +7,16 @@
 
 .radio-buttons {
   margin: 0 auto;
-}
-
-.radio-buttons.disabled {
-  pointer-events: none;
-  box-shadow: none;
-  color: $button-disabled-color !important;
-  cursor: default;
-
-  label:after {
-    background-color: $button-disabled-color;
-    border: 2px solid $button-disabled-color;
+  &.disabled {
+    pointer-events: none;
+    box-shadow: none;
+    color: $button-disabled-color !important;
+    cursor: default;
+  
+    label:after {
+      background-color: $button-disabled-color;
+      border: 2px solid $button-disabled-color;
+    }
   }
 }
+
